+++ src/core/log/engines/index.ts
@@ -6,7 +6,7 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
-import { log as console } from 'core/log/engines/console-log';
+import { log as console } from 'core/log/engines/console';
 export * from 'core/log/engines/types';
 
 const enginesStrategy = {
